update firmware esp32 version 1.5.13

+++ program_server_1/program_platformio/program_esp32_v1.5-rev3/src/10_serialData.cpp
@@ -6,7 +6,7 @@
 #include <ArduinoJson.h>
 #include "1_configProgram.h"
 
-int callback_1, callback_4;
+byte callback_1, callback_4;
 void sendDataESP(void) {
     // buat object DynamicJsonDocument data
     DynamicJsonDocument data(500);
